extract shared multi series domain helper
