+++ test/client/spec/components/victory-bar/helper-methods.spec.js
@@ -3,91 +3,8 @@
 
 import Helpers from "src/components/victory-bar/helper-methods";
 import Scale from "src/helpers/scale";
-import Domain from "src/helpers/domain";
 
 describe("victory-bar/helper-methods", () => {
-  describe("getDomain", () => {
-    let sandbox;
-    beforeEach(() => {
-      sandbox = sinon.sandbox.create();
-      sandbox.spy(Domain, "getDomainFromGroupedData");
-    });
-
-    afterEach(() => {
-      sandbox.restore();
-    });
-    const data = [
-      [{x: 1, y: 0}, {x: 2, y: 0}, {x: 3, y: 0}],
-      [{x: 1, y: 1}, {x: 2, y: 1}, {x: 3, y: 1}],
-      [{x: 1, y: 2}, {x: 2, y: 2}, {x: 3, y: 2}]
-    ];
-    const noZero = [
-      [{x: 1, y: 1}, {x: 2, y: 1}, {x: 3, y: 1}],
-      [{x: 1, y: 2}, {x: 2, y: 2}, {x: 3, y: 2}],
-      [{x: 1, y: 3}, {x: 2, y: 3}, {x: 3, y: 3}]
-    ];
-    const mixedData = [
-      [{x: 1, y: 1}, {x: 2, y: 1}, {x: 3, y: 1}],
-      [{x: 1, y: 3}, {x: 2, y: 3}, {x: 3, y: 3}],
-      [{x: 1, y: -1}, {x: 2, y: -1}, {x: 3, y: -1}],
-      [{x: 1, y: -2}, {x: 2, y: -2}, {x: 3, y: -2}]
-    ];
-
-    it("calculates a domain from props", () => {
-      const props = {domain: {x: [1, 2], y: [2, 3]}};
-      const domainResultX = Helpers.getDomain(props, "x");
-      expect(Domain.getDomainFromGroupedData).notCalled;
-      expect(domainResultX).to.eql([1, 2]);
-      const domainResultY = Helpers.getDomain(props, "y");
-      expect(Domain.getDomainFromGroupedData).notCalled;
-      expect(domainResultY).to.eql([2, 3]);
-    });
-
-    it("calculates a domain from data", () => {
-      const props = {data, grouped: true, x: "x", y: "y"};
-      const domainResultX = Helpers.getDomain(props, "x");
-      expect(Domain.getDomainFromGroupedData).calledWith(props, "x")
-        .and.returned([1, 3]);
-      expect(domainResultX).to.eql([1, 3]);
-      const domainResultY = Helpers.getDomain(props, "y");
-      expect(Domain.getDomainFromGroupedData).calledWith(props, "y")
-        .and.returned([0, 2]);
-      expect(domainResultY).to.eql([0, 2]);
-    });
-
-    it("calculates a domain from data with negative and positive values", () => {
-      const props = {data: mixedData, grouped: true, x: "x", y: "y"};
-      const domainResultX = Helpers.getDomain(props, "x");
-      expect(domainResultX).to.eql([1, 3]);
-      const domainResultY = Helpers.getDomain(props, "y");
-      expect(domainResultY).to.eql([-2, 3]);
-    });
-
-    it("the domain of the dependent axis should always contain zero", () => {
-      const props = {data: noZero, grouped: true, x: "x", y: "y"};
-      const domainResultX = Helpers.getDomain(props, "x");
-      expect(domainResultX).to.eql([1, 3]);
-      const domainResultY = Helpers.getDomain(props, "y");
-      expect(domainResultY).to.eql([0, 3]);
-    });
-
-    it("calculates a stacked domain", () => {
-      const props = {data, stacked: true, x: "x", y: "y"};
-      const domainResultX = Helpers.getDomain(props, "x");
-      expect(domainResultX).to.eql([1, 3]);
-      const domainResultY = Helpers.getDomain(props, "y");
-      expect(domainResultY).to.eql([0, 3]);
-    });
-
-    it("calculates a stacked domain from data with negative and positive values", () => {
-      const props = {data: mixedData, stacked: true, x: "x", y: "y"};
-      const domainResultX = Helpers.getDomain(props, "x");
-      expect(domainResultX).to.eql([1, 3]);
-      const domainResultY = Helpers.getDomain(props, "y");
-      expect(domainResultY).to.eql([-3, 4]);
-    });
-  });
-
   describe("getBarPosition", () => {
     let sandbox;
     beforeEach(() => {
